backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
pdf2image
google-generativeai
supabase
boto3
selenium>=4.9.0
webdriver-manager>=3.8.6
Django>=4.0.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
supabase>=1.0.0,<2.0.0
selenium>=4.10.0,<5.0.0
celery>=5.3.0,<6.0.0
redis>=4.5.0,<5.0.0
django-celery-beat>=2.5.0,<3.0.0
django-celery-results>=2.4.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
gunicorn>=21.0.0,<22.0.0
openai